// src/store/quantity/types.ts
export const Quantity = {
  isLoading: false,
  chartData: {
    title: '',
    data: [],
  },
};

export enum QuantityActions {
  ORDER_DELIVERY_REQUEST = '@@QUANTITY/ORDER_DELIVERY_REQUEST',
  ORDER_DELIVERY_SUCCESS = '@@QUANTITY/ORDER_DELIVERY_SUCCESS',
  ORDER_DELIVERY_FAILURE = '@@QUANTITY/ORDER_DELIVERY_FAILURE',

  ACTIVE_TRANSPORTER_REQUEST = '@@QUANTITY/ACTIVE_TRANSPORTER_REQUEST',
  ACTIVE_TRANSPORTER_SUCCESS = '@@QUANTITY/ACTIVE_TRANSPORTER_SUCCESS',
  ACTIVE_TRANSPORTER_FAILURE = '@@QUANTITY/ACTIVE_TRANSPORTER_FAILURE',

  AVERAGE_VOLUME_REQUEST = '@@QUANTITY/AVERAGE_VOLUME_REQUEST',
  AVERAGE_VOLUME_SUCCESS = '@@QUANTITY/AVERAGE_VOLUME_SUCCESS',
  AVERAGE_VOLUME_FAILURE = '@@QUANTITY/AVERAGE_VOLUME_FAILURE',

  AVERAGE_ORDER_TRANSPORTER_REQUEST = '@@QUANTITY/AVERAGE_ORDER_TRANSPORTER_REQUEST',
  AVERAGE_ORDER_TRANSPORTER_SUCCESS = '@@QUANTITY/AVERAGE_ORDER_TRANSPORTER_SUCCESS',
  AVERAGE_ORDER_TRANSPORTER_FAILURE = '@@QUANTITY/AVERAGE_ORDER_TRANSPORTER_FAILURE',

  DAYS_TRIP_REQUEST = '@@QUANTITY/DAYS_TRIP_REQUEST',
  DAYS_TRIP_SUCCESS = '@@QUANTITY/DAYS_TRIP_SUCCESS',
  DAYS_TRIP_FAILURE = '@@QUANTITY/DAYS_TRIP_FAILURE',

  ACTIVE_TRUCK_REQUEST = '@@QUANTITY/ACTIVE_TRUCK_REQUEST',
  ACTIVE_TRUCK_SUCCESS = '@@QUANTITY/ACTIVE_TRUCK_SUCCESS',
  ACTIVE_TRUCK_FAILURE = '@@QUANTITY/ACTIVE_TRUCK_FAILURE',

  AVERAGE_WEIGHT_REQUEST = '@@QUANTITY/AVERAGE_WEIGHT_REQUEST',
  AVERAGE_WEIGHT_SUCCESS = '@@QUANTITY/AVERAGE_WEIGHT_SUCCESS',
  AVERAGE_WEIGHT_FAILURE = '@@QUANTITY/AVERAGE_WEIGHT_FAILURE',

  AVERAGE_ORDER_TRUCK_REQUEST = '@@QUANTITY/AVERAGE_ORDER_TRUCK_REQUEST',
  AVERAGE_ORDER_TRUCK_SUCCESS = '@@QUANTITY/AVERAGE_ORDER_TRUCK_SUCCESS',
  AVERAGE_ORDER_TRUCK_FAILURE = '@@QUANTITY/AVERAGE_ORDER_TRUCK_FAILURE',

  TRIP_MONTH_REQUEST = '@@QUANTITY/TRIP_MONTH_REQUEST',
  TRIP_MONTH_SUCCESS = '@@QUANTITY/TRIP_MONTH_SUCCESS',
  TRIP_MONTH_FAILURE = '@@QUANTITY/TRIP_MONTH_FAILURE',

  KM_TRIP_REQUEST = '@@QUANTITY/_REQUEST',
  KM_TRIP_SUCCESS = '@@QUANTITY/_SUCCESS',
  KM_TRIP_FAILURE = '@@QUANTITY/_FAILURE',
}